refactor: split the leaf modules out of bridge.mjs, and test that it boots (#154)
STAGE ONE of the split, not the whole thing. Four modules come out —
log, render, stores, fanout — chosen because each has ZERO coupling to
config or ambient state, so the seam is provable rather than argued.
bridge.mjs 1748 -> 1610.

Deliberately NOT attempted here: extracting config/PUB. That is the real
work (80 PUB references as ambient module state) and it is the risky part;
doing it in the same PR as four mechanical lifts would make neither
reviewable. Issue #154 stays open for it.

The suites were NOT edited. Everything the tests import from bridge.mjs is
re-exported, so the test contract is unchanged — which is the whole point:
a split that needs a suite touched is a rewrite, not a split.

The gap this closes matters more than the line count. Every other suite
sets WB_NO_BOOT=1 — correctly, since they drive exported functions and
must not open sockets — with the consequence that until now THE BOOT BLOCK
WAS NEVER EXECUTED BY CI AT ALL. Module-level initialisation, store load
order, and config resolution were entirely unexercised. That is precisely
what a module split breaks, and verify-deployed.sh would not catch it: it
compares file hashes, and a tree that matches git perfectly can still be
dead on arrival.

tests/boot.mjs spawns the real entry point as a child process and asserts
it initialises and exits cleanly. Two negative controls, both run:
an unresolvable import fails it, and a module that resolves but is missing
a named export fails it.

The second control did not fire on its first attempt — my perl targeted
`export { defuseRefs`, but render.mjs uses inline `export const`, so it
matched nothing, exited 0, and the fallback never ran. The suite was fine;
my test of the test was broken. Re-done against the real declaration, it
fails as it should.

Also verified, because it is the failure that would only appear on the
box: all four new files are tracked and inside the deploy ship list's `src`
expansion. A deploy shipping bridge.mjs without log.mjs would not boot.

Suite count 15 -> 16, per #138's rule that package.json is the count of
record.
