+++ README.md
@@ -104,10 +104,10 @@ against what git says, because a stale build is invisible while every status sur
 
 ## Tests
 
-`npm test` runs 15 suites against the **real** exported functions with synthetic events — no
+`npm test` runs 16 suites against the **real** exported functions with synthetic events — no
 sockets, no production state:
 
-durable dedup store · relay fan-out · quarantine gating · deletion propagation · sealed-lane rate caps · grant
+boot · durable dedup store · relay fan-out · quarantine gating · deletion propagation · sealed-lane rate caps · grant
 admission · message rendering · deployed-build verification · return lane · return-lane scan ·
 return-lane no-miss · relay ingress · tripwire union · tripwire detection drill · deploy runner
